// packages/dnb-eufemia/src/extensions/forms/types.d.ts
export type DotNotationErrorMessages = Record<`${string}` | `${string}.${string}`, string | React.ReactElement>;
/**
 * { 'nb-NO': { 'Field.errorRequired': 'Dette feltet er påkrevd' } }
 */
export type ErrorMessagesWithLocaleSupport = Record<FormsTranslationLocale, DefaultErrorMessages>;
export type InternalErrorMessages = Record<Exclude<FormsTranslationFlat, undefined>, string>;
export type DefaultErrorMessages = Partial<InternalErrorMessages> & Partial<DotNotationErrorMessages>;
// ...
type EventArgs<Value, ExtraValue extends ProvideAdditionalEventArgs> = [
    value: Value,
    additionalArgs?: ExtraValue & ReceiveAdditionalEventArgs<Value>
];
export interface DataValueWriteProps<Value = unknown, EmptyValue = undefined | unknown, ExtraValue extends ProvideAdditionalEventArgs = ProvideAdditionalEventArgs> {
    emptyValue?: EmptyValue;
    onFocus?: (...args: EventArgs<Value | EmptyValue, ExtraValue>) => void;
    onBlur?: (...args: EventArgs<Value | EmptyValue, ExtraValue>) => void;
    onChange?: (...args: EventArgs<Value | EmptyValue, ExtraValue>) => OnChangeReturnType;
}
// ...
export type ProvideAdditionalEventArgs = Record<string, unknown>;
// ...
export interface UseFieldProps<Value = unknown, EmptyValue = undefined | unknown, ErrorMessages extends DefaultErrorMessages = DefaultErrorMessages, ExtraValue extends ProvideAdditionalEventArgs = ProvideAdditionalEventArgs> extends DataValueReadWriteComponentProps<Value, EmptyValue>, AriaAttributes {
    /**
     * ID added to the actual field component, and linked to the label via for-attribute
     */
    id?: Identifier;
    name?: string;
    disabled?: boolean;
    readOnly?: boolean;
    autoComplete?: HTMLInputElement['autocomplete'] | HTMLTextAreaElement['autocomplete'];
    /**
     * Text showing in place of the value if no value is given
     */
    placeholder?: React.ReactNode;
    /**
     * NB: Undocumented for now.
     * Forwards all possible props to the underlying component.
     */
    htmlAttributes?: Record<string, unknown>;
    /**
     * NB: Undocumented for now.
     * Forwards all given props in a props object.
     */
    props?: Record<string, unknown>;
    info?: InfoProp<Value>;
    warning?: WarningProp<Value>;
    error?: ErrorProp<Value>;
    /**
     * Called when the current status (info, warning or error) changes.
     * Provides the latest values for each status key.
     */
    onStatusChange?: (status: FieldStatus) => void;
    required?: boolean;
    schema?: Schema<Value> | ((props: UseFieldProps<Value>) => Schema<Value>);
    onChangeValidator?: Validator<Value>;
    onBlurValidator?: Validator<Value>;
    exportValidators?: Record<string, Validator<Value>>;
    validateRequired?: (internal: Value, { emptyValue, required, isChanged, error, }: {
        emptyValue: EmptyValue;
        required: boolean;
        isChanged: boolean;
        error: FormError | undefined;
    }) => FormError | undefined;
    /**
     * Should error messages based on validation be shown initially (from given value-prop or source data)
     * before the user interacts with the field?
     * @default false
     */
    validateInitially?: boolean;
    /**
     * Should error messages be shown when touching (like focusing a field and blurring) without having changed
     * the value? So the user did not introduce a new error, but it was invalid based on validation initially.
     */
    validateUnchanged?: boolean;
    /**
     * Should validation be done while writing, not just when blurring the field?
     */
    validateContinuously?: boolean;
    /**
     * Provide custom error messages for the field
     */
    errorMessages?: ErrorMessages;
    /**
     * Transforms the `value` before it's displayed in the field (e.g. input).
     * Public API. Should not be used internally.
     */
    transformIn?: (external: unknown) => Value | ExtraValue;
    /**
     * Transforms the value before it gets forwarded to the form data object or returned as the onChange value parameter.
     * Public API. Should not be used internally.
     */
    transformOut?: (internal: Value | ExtraValue, additionalArgs?: unknown) => unknown;
    /**
     * Transforms the value given by `handleChange` after `fromInput` and before `updateValue` and `toEvent`. The second parameter returns the current value.
     */
    transformValue?: (value: Value, currentValue?: Value) => Value;
    /**
     * Transform additionalArgs or generate it based on `value`.
     */
    provideAdditionalArgs?: (value: Value, additionalArgs?: ProvideAdditionalEventArgs) => ProvideAdditionalEventArgs;
    /**
     * Transforms the value before it gets returned as the `value`.
     */
    toInput?: (external: Value | unknown) => Value | unknown;
    /**
     * Transforms the internal value before it gets returned by even callbacks such as `onChange`, `onFocus` and `onBlur`. The second parameter returns the event type: `onChange`, `onFocus`, `onBlur` or `onBlurValidator`.
     */
    toEvent?: (internal: Value, type: 'onChange' | 'onFocus' | 'onBlur' | 'onBlurValidator') => Value;
    /**
     * Transforms the value given by `handleChange` before it is used in the further process flow. Use it to destruct the value from the original event object.
     */
    fromInput?: (external: Value | unknown) => Value;
    /**
     * Transforms the given props `value` before any other step gets entered.
     */
    fromExternal?: (external: Value) => Value;
    /**
     * For internal use only.
     */
    valueType?: string | number | boolean | Array<string | number | boolean>;
}
export type FieldProps<Value = unknown, EmptyValue = undefined | unknown, ErrorMessages extends DefaultErrorMessages = DefaultErrorMessages, ExtraValue extends ProvideAdditionalEventArgs = ProvideAdditionalEventArgs> = UseFieldProps<Value, EmptyValue, ErrorMessages, ExtraValue> & SharedFieldBlockProps;
export type FieldPropsGeneric<Value = unknown, EmptyValue = undefined | unknown, ErrorMessages extends DefaultErrorMessages = DefaultErrorMessages> = Omit<FieldProps<Value, EmptyValue, ErrorMessages>, keyof DataValueWriteProps> & DataValueWriteProps<Value, EmptyValue, ProvideAdditionalEventArgs>;
export type FieldPropsWithExtraValue<Value = unknown, ExtraValue extends ProvideAdditionalEventArgs = ProvideAdditionalEventArgs, EmptyValue = undefined | unknown, ErrorMessages extends DefaultErrorMessages = DefaultErrorMessages> = Omit<FieldProps<Value, EmptyValue, ErrorMessages, ExtraValue>, keyof DataValueWriteProps> & DataValueWriteProps<Value, EmptyValue, ExtraValue>;
// ...
type OnChangeReturnType = EventReturnWithStateObjectAndSuccess | void | Promise<EventReturnWithStateObjectAndSuccess | void>;